mads/fiedler/pyproject.toml
[project]
name = "fiedler"
version = "1.0.0"
description = "Fiedler MCP Server - Orchestra Conductor Prototype"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "openai>=1.37.0",
    "mcp>=1.2.0",
    "keyring>=24.0.0",
    "websockets>=12.0",
    "aiohttp>=3.8.0",
]

[project.scripts]
fiedler = "fiedler.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fiedler*"]

[tool.setuptools.package-data]
fiedler = ["config/models.yaml"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
